services/agent-worker/requirements.txt
asyncpg==0.30.0
nats-py==2.9.0
prometheus-client==0.21.1
pydantic==2.9.2
redis==5.1.1
langgraph>=0.2,<0.3
pytest==8.3.3

